se mejora la interfaz grafica y mejoras en flujo de uso del usuario
