+++ app/src/main/java/com/umg/todohome/taskFragment.kt
@@ -7,12 +7,14 @@ import android.content.Intent
 import android.content.pm.PackageManager
 import android.location.Location
 import android.os.Bundle
+import android.os.Handler
 import android.os.Looper
 import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.LinearLayout
 import android.widget.Toast
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
@@ -58,6 +60,16 @@ class taskFragment : Fragment() {
             val intent = Intent(requireActivity(), ActivityAddTask::class.java)
             startActivity(intent)
         }
+
+        Handler(Looper.getMainLooper()).postDelayed({
+
+            var viewRecycle = view.findViewById<LinearLayout>(R.id.viewDataTask)
+            var loading = view.findViewById<LinearLayout>(R.id.loadingCardTask)
+
+            viewRecycle.visibility = View.VISIBLE
+            loading.visibility = View.GONE
+
+        }, 500)
 
 
     }
