+++ app/src/main/res/layout/activity_data_user.xml
@@ -63,23 +63,28 @@
             android:layout_gravity="center_horizontal">
 
             <Button
-                android:layout_width="120dp"
-                android:layout_height="wrap_content"
+                android:id="@+id/saveButton"
+                android:layout_width="wrap_content"
+                android:layout_height="40dp"
                 android:background="@drawable/rounded_logo_background"
                 android:onClick="SavePhotos"
-                android:text="Subir Foto" />
+                android:text="Subir Foto"
+                android:textSize="12sp"  />
 
             <Button
-                android:layout_width="120dp"
-                android:layout_height="wrap_content"
+                android:id="@+id/deleteButton"
+                android:layout_width="wrap_content"
+                android:layout_height="40dp"
                 android:layout_marginStart="7dp"
                 android:background="@drawable/rounded_logo_background"
                 android:onClick="DeletePhoto"
-                android:text="Eliminar Foto" />
+                android:text="Eliminar Foto"
+                android:textSize="12sp"  />
 
         </LinearLayout>
 
-            <EditText
+
+        <EditText
                 android:id="@+id/user_Name"
                 android:layout_marginTop="40dp"
                 android:layout_width="match_parent"
