+++ app/src/main/res/layout/activity_main.xml
@@ -29,7 +29,9 @@
                 android:layout_height="665dp"
                 android:backgroundTint="@color/white"
                 app:layout_constraintBottom_toBottomOf="@id/bottom_navigation"
-                android:layout_marginTop="?attr/actionBarSize"/>
+                android:layout_marginTop="?attr/actionBarSize">
+
+            </FrameLayout>
 
             <com.google.android.material.bottomnavigation.BottomNavigationView
                 android:id="@+id/bottom_navigation"
@@ -40,6 +42,14 @@
                 app:itemTextColor="@color/background_highlight"
                 app:menu="@menu/nav_menu"
                 android:layout_gravity="bottom"  />
+
+            <ProgressBar
+                android:id="@+id/progressBar"
+                style="?android:attr/progressBarStyle"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                app:layout_anchor="@+id/frame_container"
+                app:layout_anchorGravity="center" />
 
         </androidx.coordinatorlayout.widget.CoordinatorLayout>
