+++ app/src/main/res/layout/card_loading.xml
@@ -1,0 +1,43 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:id="@+id/card_location"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:backgroundTint="@color/white">
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="40dp"
+        android:layout_marginTop="15dp">
+
+        <FrameLayout
+            android:layout_width="40dp"
+            android:layout_height="40dp"
+            android:background="#c8c8c9"
+            android:layout_gravity="center_horizontal"
+            android:layout_marginStart="20dp"
+            android:src="@drawable/user_image"
+            app:shapeAppearanceOverlay="@style/circular" />
+
+        <FrameLayout
+            android:layout_width="220dp"
+            android:layout_height="10dp"
+            android:layout_marginStart="8dp"
+            android:background="#c8c8c9"
+            android:layout_gravity="center_vertical"
+            android:textSize="15sp"
+            android:textColor="@color/background_highlight"
+            />
+        <FrameLayout
+            android:layout_width="match_parent"
+            android:layout_height="35dp"
+            android:layout_marginLeft="15dp"
+            android:backgroundTint="#c8c8c9"
+            android:layout_marginRight="15dp"
+            android:background="@drawable/rounded_logo_background"
+            android:layout_gravity="center_vertical"
+            android:textSize="7sp"/>
+
+    </LinearLayout>
+</androidx.cardview.widget.CardView>
