+++ app/src/main/res/layout/card_loading_expensive.xml
@@ -1,0 +1,99 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:backgroundTint="@color/white">
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="5dp"
+        android:background="@drawable/rounded_field">
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="100dp"
+            android:layout_marginTop="5dp"
+            android:orientation="vertical">
+
+            <LinearLayout
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:orientation="vertical"
+                android:layout_gravity="start">
+
+                <FrameLayout
+                    android:layout_width="220dp"
+                    android:layout_height="10dp"
+                    android:layout_marginStart="25dp"
+                    android:background="#c8c8c9"
+                    android:textSize="14sp" />
+
+                <FrameLayout
+                    android:layout_width="220dp"
+                    android:layout_height="10dp"
+                    android:layout_gravity="start"
+                    android:layout_marginTop="3dp"
+                    android:layout_marginStart="35dp"
+                    android:background="#c8c8c9"
+                    android:textSize="14sp" />
+
+            </LinearLayout>
+
+            <FrameLayout
+                android:layout_width="240dp"
+                android:layout_height="30dp"
+                android:layout_marginTop="3dp"
+                android:layout_gravity="center_vertical"
+                android:background="#c8c8c9"
+                android:layout_marginStart="35dp"
+                android:textColor="@color/background_highlight"
+                android:textSize="12sp" />
+
+            <LinearLayout
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="2dp"
+                android:orientation="horizontal">
+
+                <FrameLayout
+                    android:layout_width="40dp"
+                    android:layout_height="10dp"
+                    android:layout_marginStart="35dp"
+                    android:background="#c8c8c9"/>
+
+                <FrameLayout
+                    android:layout_width="40dp"
+                    android:layout_height="10dp"
+                    android:layout_marginStart="5dp"
+                    android:background="#c8c8c9"/>
+
+                <FrameLayout
+                    android:layout_width="70dp"
+                    android:layout_height="10dp"
+                    android:layout_marginStart="5dp"
+                    android:background="#c8c8c9"/>
+
+            </LinearLayout>
+
+        </LinearLayout>
+    </LinearLayout>
+
+    <RelativeLayout
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="right"
+        android:layout_marginTop="35dp"
+        android:paddingStart="9dp"
+        android:paddingEnd="20dp">
+
+        <FrameLayout
+            android:layout_width="70dp"
+            android:layout_height="50dp"
+            android:layout_marginStart="5dp"
+            android:background="#c8c8c9"/>
+
+    </RelativeLayout>
+
+</androidx.cardview.widget.CardView>
