+++ app/src/main/res/layout/card_task.xml
@@ -35,7 +35,7 @@
 
         <TextView
             android:id="@+id/txdescriptask"
-            android:layout_width="280dp"
+            android:layout_width="260dp"
             android:layout_height="38dp"
             android:layout_marginTop="3dp"
             android:layout_gravity="center_vertical"
